+++ internal/client/client.go
@@ -27,7 +27,7 @@ import (
 )
 
 var (
-	Version   = "1.3.2"
+	Version   = "1.3.3"
 	BuildTime = "unknown"
 
 	// Global metrics instance (nil if metrics disabled)